// extensions/arc/src/controller/generated/v1/model/hybridSqlNsNameGetResponse.ts
/**
 * SQL Server Big Data Cluster API
 * OpenAPI specification for **SQL Server Big Data Cluster**.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class HybridSqlNsNameGetResponse {
    'name'?: string;
    'status'?: string;
    'cluster_endpoint'?: string;
    'external_endpoint'?: string;
    'service_name'?: string;
    'vcores'?: string;
}